refactor(commit-tab): extract relativeTime helper; lock CommitTabState.title
- Move duplicated relativeTime(_:) out of CommitRow and CommitHeaderView
  into a shared free function in Alas/Sources/UI/RelativeTime.swift,
  matching the ChangeSummaryVisibility pattern.
- CommitTabState.title is never mutated anywhere; flip var to let.
